requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Dance==7.0.0
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==2.3.7
email-validator==2.0.0
Pillow==10.0.1
paystack==1.5.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
pyotp==2.9.0
qrcode[pil]==7.4.2
twilio==8.10.0
cryptography==41.0.7
flask-socketio==5.3.6
python-socketio==5.9.0